# app/static/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Mechanic Shop API v1"
  description: "Auto-generated minimal Swagger definitions and paths for all routes"
basePath: "/"
schemes:
  - "http"
definitions:
  Customer:
    type: object
    required: ["name", "email", "phone", "password"]
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      password:
        type: string
  CustomerUpdate:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      password:
        type: string
...
  Login:
    type: object
    required: ["email", "password"]
    properties:
      email:
        type: string
      password:
        type: string
paths:
  /customers:
    post:
      summary: Create Customer
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/Customer"
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/Customer"
        "400":
          description: Validation error
    get:
      summary: Get All Customers
      responses:
        "200":
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/Customer"
  /customers/login:
    post:
      summary: Customer Login
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/Login"
      responses:
        "200":
          description: Token returned
          schema:
            type: object
            properties:
              token:
                type: string
        "400":
          description: Bad request
  /customers/{customer_id}:
    get:
      summary: Get Customer by ID
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/Customer"
        "404":
          description: Not found
    put:
      summary: Update Customer
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CustomerUpdate"
      responses:
        "200":
          description: Updated
          schema:
            $ref: "#/definitions/Customer"
        "403":
          description: Unauthorized
    delete:
      summary: Delete Customer
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
      responses:
        "200":
          description: Deleted
        "404":
          description: Not found
...